fix(calc): sign-extend hex literals and add FMOD
Two crate bugs caught by the EPICS Base conformance corpus (Task 9):

- Hex literals were parsed as u64 and widened unsigned, so 0xffffffff
  evaluated to 4294967295. Base parses with epicsParseUInt32, stores the
  raw bits under LITERAL_INT (postfix.c:280-290) and reloads them as
  epicsInt32 (calcPerform.c:68-71), making bit 31 a sign bit: 0xffffffff
  is -1. Values wider than 32 bits are now rejected as BadNumber,
  matching Base'"'"'s CALC_ERR_BAD_LITERAL.

- FMOD was missing entirely. Base has {"FMOD", 7, 8, -1, UNARY_OPERATOR,
  FMOD} in postfix.c'"'"'s operands[] and *ptop = fmod(*ptop, top) in
  calcPerform.c. Added Op::Fmod (arity 2, function precedence) with
  argument order NOT reversed, unlike ATAN2.

Corpus failures: 10 -> 2 (both remaining are error-category ordering).

+++ spvirit-calc/src/lex.rs
@@ -87,8 +87,16 @@ fn lex_number(s: &str) -> Option<(f64, usize)> {
         if len == 0 {
             return None;
         }
-        let value = u64::from_str_radix(&rest[..len], 16).ok()?;
-        return Some((value as f64, 2 + len));
+        // Base parses hex with `epicsParseUInt32` and stores the result into
+        // the postfix stream as raw `epicsUInt32` bits under a `LITERAL_INT`
+        // tag (`refs/postfix.c:280-290`); `calcPerform` then reloads those
+        // bits as an `epicsInt32` before widening to double
+        // (`refs/calcPerform.c:68-71`). So bit 31 is a sign bit: `0xffffffff`
+        // is -1, not 4294967295. Values that do not fit in 32 bits are
+        // rejected (Base's `epicsParseUInt32` fails -> `CALC_ERR_BAD_LITERAL`,
+        // which maps to this crate's `BadNumber`).
+        let value = u32::from_str_radix(&rest[..len], 16).ok()?;
+        return Some((value as i32 as f64, 2 + len));
     }
 
     // Longest prefix that parses as a float. Bounded by the literal's length,
